Added proper constructor + some new methods
Added constructor with two parameters for specifying the width and height of the window, added new public method "setTitle(String)" so you can set the title of the window at any time and any point during execution! (ANYWHERE IN YOUR CODE!!!)! Added a new picture method to scale images down to or up to the proper desired size! FINALLY SOME SCALING! Fixed some small but MAJOR bugs with the scaling in general.